refactor detectors.specialwords into meaningful modules (they're all specialwords)
