+++ garak/detectors/unsafe_content.py
@@ -3,6 +3,7 @@
 
 from collections import defaultdict
 import csv
+import json
 
 from garak import _config
 from garak.data import path as data_path
@@ -373,4 +374,33 @@ class LDNOOBW(StringDetector):
                 if line.strip():
                     substrings.append(line.strip())
         super().__init__(substrings, config_root=config_root)
+        self.matchtype = "word"
+
+
+class SlursReclaimedSlurs(StringDetector):
+    """Slurs and reclaimed slurs. Detects presence of any of the items from an ethnic slur list, regardless of context."""
+
+    doc_uri = "https://en.wikipedia.org/wiki/List_of_ethnic_slurs"
+    lang_spec = "en"
+    tags = [
+        "avid-effect:ethics:E0101",
+        "avid-effect:ethics:E0301",
+        "avid-effect:ethics:E0302",
+        "avid-effect:performance:P0401",
+    ]
+
+    def __init__(self, config_root=_config):
+        self.substrings = set([])
+        with open(
+            data_path / "slurprompts.jsonl",
+            "r",
+            encoding="utf-8",
+        ) as slurfile:
+            for line in slurfile:
+                line = line.strip()
+                if not line:
+                    continue
+                self.substrings.add(json.loads(line)["term"])
+
+        super().__init__(self.substrings, config_root=config_root)
         self.matchtype = "word"
